data_importer/requirements.txt
apscheduler==3.3.1
arrow==0.10.0
argparse_actions==0.4.4
blinker==1.4
chardet==3.0.3
demjson==2.2.4
dicttoxml==1.7.4
flake8==3.2.1
futures==3.0.5
gevent==1.2.2
grequests==0.3.0
jinja2==2.9.5
markdown==2.6.6
pandas==0.21.0
peewee==2.8.5
plumbum==1.6.2
psutil==5.0.1
pyfunctional==1.0.0
pygments==2.1.3
pymysql==0.7.9
python-dateutil==2.6.0
python-gitlab==0.17
pathlib==1.0.1
PyYAML==3.12
redis==2.10.5
requests==2.17.3
retrying==1.3.3
six==1.10.0
toolz==0.8.2
pytest==3.2.0
pytest_allure_adaptor==1.7.8
wrapcache==1.0.8
wrapt==1.10.8
xmltodict==0.10.2
backports.functools_partialmethod==3.5.1.0
beautifulsoup4==4.5.3
PyStaticConfiguration==0.10.3
httpretty==0.8.14
fabric==1.14.0
fabtools==0.20.0
voluptuous==0.10.5
cuisine==0.7.13
Pillow==4.2.1
jmespath==0.9.2
logzero==1.3.1
jsonobject==0.7.1
xlrd==1.0.0
numpy==1.13.3